#1289: Parse results / Source Metadata Augmentation (#1810)
Add the ability to capture supplemental metadata to attach to SOURCE files. The only reason for this is to support augmentation of metadata associated with files that domain specific jobs can provide - the Scale Ingest doesn't have the smarts for it, nor should it.

We are expanding on the use of the metadata sidecar file for input files. There are some small differences. Since we can't guarantee writable storage within the input location, we must write them to the output directory. Details on the format and location of the metadata json can be found in the documentation here: https://github.com/ngageoint/scale/wiki/Scale-Job-Inputs-and-Outputs#source-input-metadata

The reason we are using the input file name as opposed to the same name as the input file name is because we can easily allow the algorithm developers at build time to know the location, since it is defined on the interface.

We do not support making input data updates to multiple file inputs. We also limit updates to SOURCE file_types.

Includes the addition of postman tests and a new Seed job to demonstrate source metadata capture.